+++ src/runtime/classbase.cs
@@ -302,5 +302,129 @@ namespace Python.Runtime
             Runtime.XDecref(self.tpHandle);
             self.gcHandle.Free();
         }
+
+
+        /// <summary>
+        /// Implements __getitem__ for reflected classes and value types.
+        /// </summary>
+        public static IntPtr mp_subscript(IntPtr ob, IntPtr idx)
+        {
+            IntPtr tp = Runtime.PyObject_TYPE(ob);
+            var cls = (ClassBase)GetManagedObject(tp);
+
+            if (cls.indexer == null || !cls.indexer.CanGet)
+            {
+                Exceptions.SetError(Exceptions.TypeError, "unindexable object");
+                return IntPtr.Zero;
+            }
+
+            // Arg may be a tuple in the case of an indexer with multiple
+            // parameters. If so, use it directly, else make a new tuple
+            // with the index arg (method binders expect arg tuples).
+            IntPtr args = idx;
+            var free = false;
+
+            if (!Runtime.PyTuple_Check(idx))
+            {
+                args = Runtime.PyTuple_New(1);
+                Runtime.XIncref(idx);
+                Runtime.PyTuple_SetItem(args, 0, idx);
+                free = true;
+            }
+
+            IntPtr value;
+
+            try
+            {
+                value = cls.indexer.GetItem(ob, args);
+            }
+            finally
+            {
+                if (free)
+                {
+                    Runtime.XDecref(args);
+                }
+            }
+            return value;
+        }
+
+
+        /// <summary>
+        /// Implements __setitem__ for reflected classes and value types.
+        /// </summary>
+        public static int mp_ass_subscript(IntPtr ob, IntPtr idx, IntPtr v)
+        {
+            IntPtr tp = Runtime.PyObject_TYPE(ob);
+            var cls = (ClassBase)GetManagedObject(tp);
+
+            if (cls.indexer == null || !cls.indexer.CanSet)
+            {
+                Exceptions.SetError(Exceptions.TypeError, "object doesn't support item assignment");
+                return -1;
+            }
+
+            // Arg may be a tuple in the case of an indexer with multiple
+            // parameters. If so, use it directly, else make a new tuple
+            // with the index arg (method binders expect arg tuples).
+            IntPtr args = idx;
+            var free = false;
+
+            if (!Runtime.PyTuple_Check(idx))
+            {
+                args = Runtime.PyTuple_New(1);
+                Runtime.XIncref(idx);
+                Runtime.PyTuple_SetItem(args, 0, idx);
+                free = true;
+            }
+
+            // Get the args passed in.
+            var i = Runtime.PyTuple_Size(args);
+            IntPtr defaultArgs = cls.indexer.GetDefaultArgs(args);
+            var numOfDefaultArgs = Runtime.PyTuple_Size(defaultArgs);
+            var temp = i + numOfDefaultArgs;
+            IntPtr real = Runtime.PyTuple_New(temp + 1);
+            for (var n = 0; n < i; n++)
+            {
+                IntPtr item = Runtime.PyTuple_GetItem(args, n);
+                Runtime.XIncref(item);
+                Runtime.PyTuple_SetItem(real, n, item);
+            }
+
+            // Add Default Args if needed
+            for (var n = 0; n < numOfDefaultArgs; n++)
+            {
+                IntPtr item = Runtime.PyTuple_GetItem(defaultArgs, n);
+                Runtime.XIncref(item);
+                Runtime.PyTuple_SetItem(real, n + i, item);
+            }
+            // no longer need defaultArgs
+            Runtime.XDecref(defaultArgs);
+            i = temp;
+
+            // Add value to argument list
+            Runtime.XIncref(v);
+            Runtime.PyTuple_SetItem(real, i, v);
+
+            try
+            {
+                cls.indexer.SetItem(ob, real);
+            }
+            finally
+            {
+                Runtime.XDecref(real);
+
+                if (free)
+                {
+                    Runtime.XDecref(args);
+                }
+            }
+
+            if (Exceptions.ErrorOccurred())
+            {
+                return -1;
+            }
+
+            return 0;
+        }
     }
 }
